Step 4 EDIT: The texts have been changed to English. Classes have been reordered. The recyclerView data is loaded by a method before pressing the button. Modification to the Picaso implementation. Aesthetic changes.
