+++ app/src/main/java/com/mstudio/superheroarch/Character.kt
@@ -1,0 +1,11 @@
+package com.mstudio.superheroarch
+
+data class Character(
+    val name: String,
+    val status: String,
+    val image: String
+)
+
+data class CharacterResponse(
+    val results: List<Character>
+)
