+++ app/src/main/res/values/colors.xml
@@ -2,4 +2,5 @@
 <resources>
     <color name="black">#FF000000</color>
     <color name="white">#FFFFFFFF</color>
+    <color name="off_white">#F5F5F5</color>
 </resources>
